Big commit

+++ app/src/main/java/com/redugsi/weatherapp/model/Clouds.kt
@@ -3,4 +3,4 @@ package com.redugsi.weatherapp.model
 import com.google.gson.annotations.SerializedName
 
 data class Clouds(@SerializedName("all")
-                var all: Int? = null)
+                var all: Double? = null)
