+++ app/src/main/java/com/redugsi/weatherapp/model/Main.kt
@@ -5,9 +5,9 @@ import com.google.gson.annotations.SerializedName
 class Main(@SerializedName("temp")
                var temp: Double? = null,
                @SerializedName("pressure")
-               var pressure: Int? = null,
+               var pressure: Double? = null,
                @SerializedName("humidity")
-               var humidity: Int? = null,
+               var humidity: Double? = null,
                @SerializedName("temp_min")
                var tempMin: Double? = null,
                @SerializedName("temp_max")
